+++ src/fsm/depth1/explore_for_cache_fsm.cpp
@@ -39,7 +39,7 @@ NS_START(fordyca, fsm, depth1);
  * Constructors/Destructors
  ******************************************************************************/
 explore_for_cache_fsm::explore_for_cache_fsm(
-    double unsuccessful_dir_change_thresh,
+    uint unsuccessful_dir_change_thresh,
     const std::shared_ptr<rcppsw::er::server> &server,
     const std::shared_ptr<controller::depth1::foraging_sensors> &sensors,
     const std::shared_ptr<controller::actuator_manager> &actuators)
